gh-93. Rename hook file auth -> useAuth
