+++ apps/web-portal-client/lib/features/auth/hooks/index.jsx
@@ -1,1 +1,1 @@
-export * from './auth'
+export * from './useAuth'
